Rename util/tc.py to util/taskcluster.py
